% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{attendance}
\alias{attendance}
\title{attendance}
\usage{
attendance(regular = TRUE, playoffs = TRUE, season = NULL)
}
\arguments{
\item{regular}{boolean}

\item{playoffs}{boolean}

\item{season}{string}
}
\value{
factor
}
\description{
This function will make a query to the Attendance API to get
the NHL's seasonal and playoff attendance numbers.
The function will display the NHL seasonal attendance numbers in a ggplot chart
}
\examples{
rpuck.attendance(regular=True, playoffs=True, season= "2000-11")

}
