% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{Return the statistics of player}
\usage{
player_stats(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
player_stats(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])

}
