% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{Make a sorted frequency table for a factor}
\usage{
team_stats(x)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Make a sorted frequency table for a factor
}
\examples{
team_stats(iris$Species)
}
