% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_pick.R
\name{draft_pick}
\alias{draft_pick}
\title{draft_pick}
\usage{
draft_pick(pick_number = 1, round_number = None, year = None)
}
\arguments{
\item{pick_number}{int in range [1,38]}

\item{round_number}{int in in range [1,25]}

\item{year}{str containing a year from 1963 to 2019}
}
\value{
data.frame
}
\description{
Query the result of draft by year summary report endpoint on the NHL.com API.
}
\details{
Depending on information provided in the parameters function will return summary 
on a person picked in specified draft round, pick and year. If year is not specified 
then all of the draft picks for that year will be returned. If no round is specified 
the dataframe will include all of the players with chosen pick number from every round.
}
\examples{
draft_pick(pick_number = 1, round_number = 2, year = '2019')
}
