% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{player_stats}
\usage{
player_stats(start_date = NULL, end_date = NULL)
}
\arguments{
\item{a}{string}

\item{b}{string}
}
\value{
data.frame
}
\description{
Query the top 100 player's stats (sorted by total points)
from the player's summary report endpoint on the NHL.com API.
}
\details{
The stats are queried on an aggregated game-by-game basis
for a range of dates. If no date is specified the function will
return the player's stats for the current season. The stats to be
returned are restricted to the regular season.

The function will return the current season's stats if the arguments
are blank (i.e. left as None).
}
\examples{
player_stats(start_date = "2019-10-02", end_date = "2020-02-28")
}
