% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{Team_stats}
\usage{
team_stats(start_season = NULL, end_season = NULL)
}
\arguments{
\item{start_season}{string}

\item{end_season}{string}
}
\value{
data.frame
}
\description{
Query the seasons stats for teams specified by start year or start year and end year
from the Team Summary Stats API to the the Team's Summary season stats.
}
\details{
The stats are queried on an aggregated team by team basis
for a range of dates. If no date is specified the function will
return the team stats for the current season.
The available seasons range from 1975-1976 to 2018-2019
If no year is specified then the year 2019-2020 is default.
If an end year is specified then the start year is also to be provided.
Year input is to be provided in a 2 year format of 'YYYYYYYY'
}
\examples{
team_stats(start_season = "19801981", end_season = "19891990")
}
